Change water cost input to float for accuracy

+++ app.py
@@ -1,9 +1,9 @@
 tipo = input("Qual seu estilo de localidade entre: casa, apartamento ou comercio?.")
 if tipo == "casa" or tipo == "apartamento":
-  custo = int(input("Quanto você gasta de água em m3?."))
-if (tipo == "casa" or tipo =="apartamento") and custo < 10:
+  custo = float(input("Quanto você gasta de água em m3?."))
+if (tipo == "casa" or tipo =="apartamento") and custo < 10.0:
   print("Consumo econômico - exelente controle de água.")
-elif (tipo == "casa" or tipo =="apartamento") and 10 <= custo <= 25:
+elif (tipo == "casa" or tipo =="apartamento") and 10.0 <= custo <= 25.0:
   print("Consumo moderado - dentro do padrão residencial.")
 elif tipo == "comercial":
   print("Comercial aplicada - procure o plano corporativo.")
